Final commit!!!
In this commit I've added one more Algorithm function for output tree on
the screen.

Today all labs was ended, and I'm proud to say that I've done with this
subject. C++ assignments in our university suck, really.

So, if you are a student of BSUIR, you can freely use this awful code in
your assignments